[Refactor] Réorganisation de la structure du projet
Ajout des sous-packages dédiés : model, dto ...
