+++ src/main/java/com/fzamiche/back_book_social_network/book/repository/BookRepository.java
@@ -1,5 +1,6 @@
-package com.fzamiche.back_book_social_network.book;
+package com.fzamiche.back_book_social_network.book.repository;
 
+import com.fzamiche.back_book_social_network.book.model.Book;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
